% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis.R
\name{remap_with_pal}
\alias{remap_with_pal}
\title{Remap a matrix with the given palette}
\usage{
remap_with_pal(mat, pal)
}
\arguments{
\item{mat}{numeric matrix}

\item{pal}{RGB palette.  256x3 numeric values in the range [0, 1]}
}
\description{
Remap a matrix with the given palette
}
