% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis.R
\name{remap_with_viridis}
\alias{remap_with_viridis}
\title{Remap a matrix with the named viridis palette}
\usage{
remap_with_viridis(mat, pal_name = "B")
}
\arguments{
\item{mat}{numeric matrix}

\item{pal_name}{viridis palette name. c('A', 'B', 'C', 'D', 'E') or
c('magma' , 'inverno', 'plasma', 'viridis', 'cividis')}
}
\description{
Remap a matrix with the named viridis palette
}
