% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm.R
\name{write_pgm}
\alias{write_pgm}
\title{Write a vector as a grayscale PGM file without sanity checking the data}
\usage{
write_pgm(values, nrow, ncol, filename)
}
\arguments{
\item{values}{Integer values preferably, but numeric values allowed.
Length should be nrow * ncol, but it is up to the caller to check.
Values are not checked for the presence of NAs.}

\item{nrow, ncol}{Dimensions of image to output}

\item{filename}{Filename to write to. Will be silently overwriten if it already exists.}
}
\description{
First pixel is at the top-left of the image, subsequent pixels are filled
along a row.
}
\details{
If \code{values} is of type \code{double}, then they will be scaled to
integers in the range [0,255]

For \code{write_pgm()} the caller is expected to have done all sanity
checking as the input is not checked for correctness, i.e.
\itemize{
\item{Not checked that \code{is.numeric(values)}}
\item{Not checked that \code{length(values) = nrow * ncol}}
\item{Not checked that all integer values in range [0, 255] - Values outside the range [0,255] will be written modulo 256}
\item{Not checked that there are no NA values}
}
}
