% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm.R
\name{write_pgm_from_matrix}
\alias{write_pgm_from_matrix}
\title{Write a 2d numeric matrix as a grayscale PGM without sanity checking the data}
\usage{
write_pgm_from_matrix(mat, filename)
}
\arguments{
\item{mat}{numeric matrix. No NAs allowed}

\item{filename}{Filename to write to. Will be silently overwriten if it already exists.}
}
\description{
Write a 2d numeric matrix as a grayscale PGM without sanity checking the data
}
