% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnm.R
\name{write_pnm}
\alias{write_pnm}
\alias{write_pnm_from_matrix}
\alias{write_pnm_from_array}
\title{Write values as a PGM/PPM file (includes checking of input data for sanity)}
\usage{
write_pnm(values, nrow, ncol, filename, nbytes = 1L, overwrite = FALSE)

write_pnm_from_matrix(values, filename, nbytes = 1L)

write_pnm_from_array(values, filename, nbytes = 1L)
}
\arguments{
\item{values}{Integer values. Length must be nrow * ncol}

\item{nrow, ncol}{Dimensions of output image}

\item{filename}{Filename to write to.}

\item{nbytes}{Number of bytes per colour compoment. Default 1L. Otherwise 2}

\item{overwrite}{Overwrite the output file if it already exists? default: FALSE}
}
\description{
Write values as a PGM/PPM file (includes checking of input data for sanity)
}
\details{
If values are double, then scaled to [0,255] or [0,65535] depending upon \code{nbytes}

\itemize{
\item{\code{write_pnm}} - write a vector of values to either a PGM or PPM file.  If
the number of values supplied exactly matches the dimensions of the image, a PGM file
is written, otherwise a PPM file is written.
\item{\code{write_pnm_matrix}} - write a 2D matrix of values as a PGM file.
\item{\code{write_pnm_array}} - write a 3D array of values as a PPM file
}
}
