% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{write_ppm}
\alias{write_ppm}
\title{Write vector as a colour PPM file without sanity checking data}
\usage{
write_ppm(values, nrow, ncol, filename)
}
\arguments{
\item{values}{Integer values. Length should be \code{3 * nrow * ncol} but
is not checked.  Values outside the range [0,255] will be written modulo 256}

\item{nrow, ncol}{Dimensions of image}

\item{filename}{Filename to write to. Will be silently overwriten if it already exists.}
}
\description{
If \code{values} is of type \code{double}, then they will be scaled to
integers in the range [0,255]
}
