% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{write_ppm_from_array}
\alias{write_ppm_from_array}
\title{Write a 3d numeric array as a colour PPM image}
\usage{
write_ppm_from_array(arr, filename)
}
\arguments{
\item{arr}{3d Array}

\item{filename}{Filename to write to. Will be silently overwriten if it already exists.}
}
\description{
Write a 3d numeric array as a colour PPM image
}
