% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{write_ppm_with_viridis}
\alias{write_ppm_with_viridis}
\title{Write numeric values as a PPM file by first scaling into [0-255] and then
using a viridis palette for colour lookup}
\usage{
write_ppm_with_viridis(values, nrow, ncol, filename, pal = "D")
}
\arguments{
\item{values}{Integer values preferably, but numeric values allowed.
Length should be nrow * ncol, but it is up to the caller to check.
Values are not checked for the presence of NAs.}

\item{nrow}{Dimensions of image to output}

\item{ncol}{Dimensions of image to output}

\item{filename}{Filename to write to. Will be silently overwriten if it already exists.}

\item{pal}{Viridis palette. Valid = c('A', 'B', 'C', 'D', 'E') or
c('magma' , 'inverno', 'plasma', 'viridis', 'cividis')}
}
\description{
Write numeric values as a PPM file by first scaling into [0-255] and then
using a viridis palette for colour lookup
}
